% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaDataTrentino}
\alias{getMetaDataTrentino}
\title{Get metedata for Trentino Weather Station}
\usage{

  getMetaDataTrentino(url = "http://dati.meteotrentino.it/service.asmx/listaStazioni",
  return.type = c("data.frame", "attributed.list", "list"),
  header = c("latitude", "longitude", "station_name",
  "station_shortname", "altitude"))
}
\arguments{
\item{url}{URL with data. Default is \code{"http://dati.meteotrentino.it/service.asmx/listaStazioni"}.}

\item{return.type}{character string indicating the requested type for the function output. Default is \code{"data.frame"}.}

\item{header}{string vectors of station metadata, used if \code{return.type=="list"}.}
}
\description{
Get metedata for Trentino Weather Station
}
\details{
The data are licensed as Open Data and released by Provincia Autonoma di Trento (\url{www.meteotrentino.it}) through \url{http://dati.trentino.it/dataset/anagrafica-stazioni-meteo-stazioni-automatiche} and \url{http://dati.trentino.it/dataset/anagrafica-campi-neve}. Please see the links for major details.
}
\examples{

# WEATHER STATIONS
url <-  "http://dati.meteotrentino.it/service.asmx/listaStazioni"
metadata <- getMetaDataTrentino(url=url)
metadata_l <- getMetaDataTrentino(url=url,return.type="attributed.list")

# SNOW MEASUREMENT STATIONS 

url_snow <- "http://dati.meteotrentino.it/service.asmx/listaCampiNeve"
metadata_snow <- getMetaDataTrentino(url=url_snow)
metadata_snow_l <- getMetaDataTrentino(url=url_snow,return.type="attributed.list")

### raw_out_data

}
\author{
Emanuele Cordano
}
