% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSnowData.R
\name{getSnowData}
\alias{getSnowData}
\title{Get data for Trentino Weather Station}
\usage{

  getSnowData(url = "http://dati.meteotrentino.it/service.asmx/tuttiUltimiRilieviNeve",
  url_metadata = "http://dati.meteotrentino.it/service.asmx/listaCampiNeve",
  header.fields = c("station_id", "station_name", "station_shortname",
  "altitude", "location", "longitude", "latitude"),
  station_id_prefix = "TRENTINO_", smet = TRUE, tz = "Etc/GMT-1", ...)
}
\arguments{
\item{url}{URL with data. Default is \code{"http://dati.meteotrentino.it/service.asmx/tuttiUltimiRilieviNeve"}.}

\item{url_metadata}{URL with data.  Default is \code{"http://dati.meteotrentino.it/service.asmx/listaCampiNeve"}.}

\item{header.fields}{header fields for the returened \code{\link{smet-class}} objects. See \code{\link{as.smet}}. It is used in case \code{smet==TRUE}.}

\item{station_id_prefix}{string added to the station ID as a prefix. It distinguishes \url{www.meteotrentino.it} snow stations from the ones of other weather services.}

\item{smet}{logical value. If it is \code{TRUE}, data are returned as a list of \code{\link{smet-class}} object(s).}

\item{tz}{time zone. Default is \code{Etc?GMT-1}.}

\item{...}{further arguments}
}
\description{
Get data for Trentino Weather Station
}
\details{
The data are licensed as Open Data and released by Provincia Autonoma di Trento (\url{www.meteotrentino.it}) through \url{http://dati.trentino.it/dataset/dati-recenti-dei-campi-neve}. Please see the link for major details.
}
\examples{


out <- getSnowData()

### LEGEND (AINEVA) : http://www.meteotrentino.it/neve-ghiacci/Husky/mod1/legenda-mod1.pdf




}
\author{
Emanuele Cordano
}

