% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeteo.R
\name{getMeteoDataTrentino}
\alias{getMeteoDataTrentino}
\title{Get data for Trentino Weather Station}
\usage{
getMeteoDataTrentino(station = c("T0179", "T0175"),
  url = "http://dati.meteotrentino.it/service.asmx/ultimiDatiStazione?codice=TCODE",
  verbose = TRUE, tz = "Etc/GMT-1", smet = TRUE, ...)
}
\arguments{
\item{station}{ID code}

\item{url}{URL with data. Default is \code{"http://dati.meteotrentino.it/service.asmx/ultimiDatiStazione?codice=T0179"}.}

\item{verbose}{verbose logical argument}

\item{tz}{time zone. Default is \code{Etc/GMT-1}.}

\item{smet}{logical value. If it is \code{TRUE}, data are returned as a list of \code{\link{smet-class}} object(s).}

\item{...}{further arguments}
}
\description{
Get data for Trentino Weather Station
}
\details{
The data are licensed as Open Data and released by Provincia Autonoma di Trento (\url{www.meteotrentino.it}) through \url{http://dati.trentino.it/dataset/dati-recenti-delle-stazioni-meteo}. Please see the link for major details.
}
\examples{


metadata <- getMetaDataTrentino(return.type="list")

nn <- names(metadata)[!(names(metadata) \%in\% c("T0365"))]

# Please uncomment the following run to run the command
# data <- getMeteoDataTrentino(station=metadata[nn])
#

datap <- getMeteoDataTrentino(station=metadata[["T0153"]])



}
\author{
Emanuele Cordano
}

